Fixes FER2-34: decompose AttemptSubmit pipeline into services with golden contract
